+++ lib/gollum/frontend/views/error.rb
@@ -1,0 +1,7 @@
+module Precious
+  module Views
+    class Error < Layout
+      attr_reader :message
+    end
+  end
+end